add mutable list class for sqlalchemy models

+++ tests/integration/test_reviews.py
@@ -22,26 +22,75 @@ def test_get_review_by_id(client, review_1):
     assert response.get_json() == review_1
 
 
-# TODO
 def test_create_review(client, mocker):
     mocker.patch("app.main.routes.reviews.get_current_user_id", return_value=1)
-    response = client.post("/api/reviews", json={})  # TODO
+    mocker.patch(
+        "app.utils.auth0.verify_decode_jwt",
+        return_value={"permissions": ["post:reviews"]},
+    )
+    response = client.post(
+        "/api/reviews",
+        json={"game_id": 5, "review_text": "Really great", "rating": 4},
+        headers={"Authorization": "Bearer x"},
+    )
     assert response.status_code == 201
-    assert response.get_json() == ""  # TODO
+    assert response.get_json() == {
+        "success": True,
+        "created": 8,
+        "reviews": [
+            {
+                "id": 8,
+                "user": 1,
+                "game_id": 5,
+                "rating": 4,
+                "review_text": "Really great",
+            }
+        ],
+        "total_reviews_for_game": 1,
+    }
 
 
-# TODO
 def test_update_review(client, mocker):
     mocker.patch("app.main.routes.reviews.check_user_id", return_value=True)
-    response = client.patch("/api/reviews/1", json={})  # TODO
+    mocker.patch(
+        "app.utils.auth0.verify_decode_jwt",
+        return_value={"permissions": ["patch:reviews"]},
+    )
+    response = client.patch(
+        "/api/reviews/1", json={"rating": 1}, headers={"Authorization": "Bearer x"}
+    )
+
     assert response.status_code == 200
-    assert response.get_json() == ""  # TODO
+    assert response.get_json() == {
+        "reviews": [
+            {
+                "game_id": 1,
+                "id": 1,
+                "rating": 1,
+                "review_text": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
+                "user": 1,
+            },
+            {
+                "game_id": 1,
+                "id": 5,
+                "rating": 5,
+                "review_text": "Id interdum velit laoreet id donec ultrices. Venenatis a condimentum vitae sapien pellentesque.",
+                "user": 2,
+            },
+        ],
+        "success": True,
+        "total_reviews_for_game": 2,
+        "updated": 1,
+    }
 
 
-# TODO
 def test_delete_review(client, mocker):
     mocker.patch("app.main.routes.reviews.check_user_id", return_value=True)
-    response = client.delete("/api/reviews/3")
+    mocker.patch(
+        "app.utils.auth0.verify_decode_jwt",
+        return_value={"permissions": ["delete:reviews"]},
+    )
+    response = client.delete("/api/reviews/3", headers={"Authorization": "Bearer x"})
     assert response.status_code == 200
     assert response.get_json() == {"success": True, "deleted": 3}
 
@@ -57,9 +106,19 @@ def test_get_reactions_for_review(client):
     }
 
 
-# TODO (auth)
-def test_like_review(client):
-    response = client.patch("/api/reviews/2/reactions", json={"like": True})
+# TODO: resolve issue with user_likes list
+def test_like_review(client, mocker):
+    mocker.patch("app.main.routes.reviews.check_user_id", return_value=True)
+    mocker.patch("app.main.routes.reviews.get_current_user_id", return_value=1)
+    mocker.patch(
+        "app.utils.auth0.verify_decode_jwt",
+        return_value={"permissions": ["patch:reactions"]},
+    )
+    response = client.patch(
+        "/api/reviews/2/reactions",
+        json={"like": True},
+        headers={"Authorization": "Bearer x"},
+    )
     assert response.status_code == 200
     assert response.get_json() == {
         "success": True,
@@ -69,9 +128,19 @@ def test_like_review(client):
     }
 
 
-# TODO (auth)
-def test_dislike_review(client):
-    response = client.patch("/api/reviews/2/reactions", json={"dislike": True})
+# TODO: resolve issue with user_dislikes list
+def test_dislike_review(client, mocker):
+    mocker.patch("app.main.routes.reviews.check_user_id", return_value=True)
+    mocker.patch("app.main.routes.reviews.get_current_user_id", return_value=1)
+    mocker.patch(
+        "app.utils.auth0.verify_decode_jwt",
+        return_value={"permissions": ["patch:reactions"]},
+    )
+    response = client.patch(
+        "/api/reviews/2/reactions",
+        json={"dislike": True},
+        headers={"Authorization": "Bearer x"},
+    )
     assert response.status_code == 200
     assert response.get_json() == {
         "success": True,
